refactor: 도메인과 vo 패키지 분리
