+++ src/main/java/kitchenpos/products/tobe/domain/Product.java
@@ -1,5 +1,8 @@
 package kitchenpos.products.tobe.domain;
 
+import kitchenpos.products.tobe.vo.DisplayedName;
+import kitchenpos.products.tobe.vo.Price;
+
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
